+++ src/lib.rs
@@ -3,8 +3,10 @@
 mod cartidge;
 mod graphics;
 mod interfaces;
+mod mappers;
 mod nes;
 mod processor;
+mod types;
 mod utils;
 
 pub use cartidge::Cartidge;